+++ anvil/utils/scene.py
@@ -14,16 +14,11 @@ def is_types(node, types):
 
 
 def safe_delete(node_or_nodes):
-    if isinstance(node_or_nodes, list):
-        for node in node_or_nodes:
-            try:
-                rt.dcc.scene.delete(node)
-            except ValueError:
-                pass
-    else:
+    node_or_nodes = to_list(node_or_nodes)
+    for node in node_or_nodes:
         try:
-            rt.dcc.scene.delete(node_or_nodes)
-        except ValueError:
+            rt.dcc.scene.delete(node)
+        except (ValueError, RuntimeError):
             pass
 
 
@@ -31,7 +26,7 @@ def objects_exist(nodes):
     return all([rt.dcc.scene.exists(node) if node else False for node in nodes])
 
 
-def list_scene_nodes(object_type=cfg.TRANSFORM_TYPE, has_shape=False):
+def list_scene_nodes(object_type=cfg.TRANSFORM_TYPE):
     nodes = []
     for node in rt.dcc.scene.list_scene(type=object_type):
         if not node.getShape():
@@ -40,10 +35,14 @@ def list_scene_nodes(object_type=cfg.TRANSFORM_TYPE, has_shape=False):
             if not node.getShape().type() == 'camera':
                 nodes.append(node)
             else:
-                if not rt.dcc.scene.API.camera(node.getShape(), startupCamera=True, q=True):
+                if not rt.dcc.ENGINE_API.camera(node.getShape(), startupCamera=True, q=True):
                     nodes.append(node)
 
     return nodes
+
+
+def sanitize_scene():
+    safe_delete(list_scene_nodes())
 
 
 def get_scene_tree():
